Rename old Jemu class to OldJemu
