+++ src/main/java/io/github/arkosammy12/jemu/application/adapters/DefaultCosmacVIPAdapter.java
@@ -1,7 +1,6 @@
 package io.github.arkosammy12.jemu.application.adapters;
 
 import io.github.arkosammy12.jemu.application.Jemu;
-import io.github.arkosammy12.jemu.application.NewJemu;
 import io.github.arkosammy12.jemu.application.io.initializers.CoreInitializer;
 import io.github.arkosammy12.jemu.application.drivers.DefaultAudioRendererDriver;
 import io.github.arkosammy12.jemu.application.drivers.JPanelVideoDriver;
@@ -12,14 +11,12 @@ import io.github.arkosammy12.jemu.backend.common.Emulator;
 import io.github.arkosammy12.jemu.backend.cosmacvip.CosmacVIPKeypad;
 import io.github.arkosammy12.jemu.backend.cosmacvip.CosmacVipEmulator;
 import io.github.arkosammy12.jemu.backend.cosmacvip.CosmacVIPHost;
-import io.github.arkosammy12.jemu.backend.drivers.AudioDriver;
 import io.github.arkosammy12.jemu.backend.drivers.VideoDriver;
 import io.github.arkosammy12.jemu.frontend.audio.AudioRenderer;
 import io.github.arkosammy12.jemu.frontend.audio.MonoAudioRenderer;
 import io.github.arkosammy12.jemu.frontend.audio.StereoAudioRenderer;
 import org.jetbrains.annotations.Nullable;
 
-import javax.swing.*;
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
 import java.io.IOException;
@@ -38,7 +35,7 @@ public class DefaultCosmacVIPAdapter extends DefaultSystemAdapter implements Cos
     private final DefaultAudioRendererDriver audioDriver;
     private final AudioRenderer audioRenderer;
 
-    public DefaultCosmacVIPAdapter(NewJemu jemu, CoreInitializer initializer, Chip8Interpreter chip8Interpreter) {
+    public DefaultCosmacVIPAdapter(Jemu jemu, CoreInitializer initializer, Chip8Interpreter chip8Interpreter) {
         super(initializer);
 
         this.romTitle = initializer.getRomPath().map(path -> path.getFileName().toString()).orElse(null);
