+++ src/main/java/io/github/arkosammy12/jemu/application/adapters/DefaultGameBoyAdapter.java
@@ -1,7 +1,6 @@
 package io.github.arkosammy12.jemu.application.adapters;
 
 import io.github.arkosammy12.jemu.application.Jemu;
-import io.github.arkosammy12.jemu.application.NewJemu;
 import io.github.arkosammy12.jemu.application.io.initializers.CoreInitializer;
 import io.github.arkosammy12.jemu.application.drivers.DefaultAudioRendererDriver;
 import io.github.arkosammy12.jemu.application.drivers.JPanelVideoDriver;
@@ -10,7 +9,6 @@ import io.github.arkosammy12.jemu.application.drivers.StereoAudioRendererDriver;
 import io.github.arkosammy12.jemu.application.util.System;
 import io.github.arkosammy12.jemu.backend.common.Emulator;
 import io.github.arkosammy12.jemu.backend.common.SystemHost;
-import io.github.arkosammy12.jemu.backend.drivers.AudioDriver;
 import io.github.arkosammy12.jemu.backend.drivers.VideoDriver;
 import io.github.arkosammy12.jemu.backend.gameboy.GameBoyEmulator;
 import io.github.arkosammy12.jemu.backend.gameboy.GameBoyHost;
@@ -21,7 +19,6 @@ import io.github.arkosammy12.jemu.frontend.audio.StereoAudioRenderer;
 import org.jetbrains.annotations.Nullable;
 import org.tinylog.Logger;
 
-import javax.swing.*;
 import java.awt.event.KeyAdapter;
 import java.awt.event.KeyEvent;
 import java.io.IOException;
@@ -42,7 +39,7 @@ public class DefaultGameBoyAdapter extends DefaultSystemAdapter implements GameB
     private final DefaultAudioRendererDriver audioDriver;
     private final AudioRenderer audioRenderer;
 
-    public DefaultGameBoyAdapter(NewJemu jemu, CoreInitializer initializer, Model model) {
+    public DefaultGameBoyAdapter(Jemu jemu, CoreInitializer initializer, Model model) {
         super(initializer);
         StringBuilder titleBuilder;
         String title = null;
