+++ src/main/java/io/github/arkosammy12/jemu/application/drivers/DefaultAudioRendererDriver.java
@@ -1,7 +1,6 @@
 package io.github.arkosammy12.jemu.application.drivers;
 
 import io.github.arkosammy12.jemu.application.Jemu;
-import io.github.arkosammy12.jemu.application.NewJemu;
 import io.github.arkosammy12.jemu.backend.common.AudioGenerator;
 import io.github.arkosammy12.jemu.backend.drivers.AudioDriver;
 import io.github.arkosammy12.jemu.frontend.audio.AudioRenderer;
@@ -13,7 +12,7 @@ public abstract class DefaultAudioRendererDriver implements AudioDriver, Closeab
     protected final AudioGenerator<?> audioGenerator;
     protected final AudioRenderer audioRenderer;
 
-    public DefaultAudioRendererDriver(NewJemu jemu, AudioGenerator<?> audioGenerator, AudioRenderer audioRenderer) {
+    public DefaultAudioRendererDriver(Jemu jemu, AudioGenerator<?> audioGenerator, AudioRenderer audioRenderer) {
         this.audioGenerator = audioGenerator;
         this.audioRenderer = audioRenderer;
         //jemu.addEmulatorFrameListener((_, _) -> this.onFrame());
