+++ src/main/java/io/github/arkosammy12/jemu/application/drivers/MonoAudioRendererDriver.java
@@ -1,6 +1,6 @@
 package io.github.arkosammy12.jemu.application.drivers;
 
-import io.github.arkosammy12.jemu.application.NewJemu;
+import io.github.arkosammy12.jemu.application.Jemu;
 import io.github.arkosammy12.jemu.backend.common.AudioGenerator;
 import io.github.arkosammy12.jemu.frontend.audio.AudioRenderer;
 import io.github.arkosammy12.jemu.frontend.audio.MonoAudioRenderer;
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 public class MonoAudioRendererDriver extends DefaultAudioRendererDriver {
 
-    public MonoAudioRendererDriver(NewJemu jemu, AudioGenerator<?> audioGenerator, MonoAudioRenderer audioRenderer) {
+    public MonoAudioRendererDriver(Jemu jemu, AudioGenerator<?> audioGenerator, MonoAudioRenderer audioRenderer) {
         super(jemu, audioGenerator, audioRenderer);
     }
