+++ src/main/java/io/github/arkosammy12/jemu/application/drivers/StereoAudioRendererDriver.java
@@ -1,6 +1,6 @@
 package io.github.arkosammy12.jemu.application.drivers;
 
-import io.github.arkosammy12.jemu.application.NewJemu;
+import io.github.arkosammy12.jemu.application.Jemu;
 import io.github.arkosammy12.jemu.backend.common.AudioGenerator;
 import io.github.arkosammy12.jemu.frontend.audio.StereoAudioRenderer;
 
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 public class StereoAudioRendererDriver extends DefaultAudioRendererDriver {
 
-    public StereoAudioRendererDriver(NewJemu jemu, AudioGenerator<?> audioGenerator, StereoAudioRenderer audioRenderer) {
+    public StereoAudioRendererDriver(Jemu jemu, AudioGenerator<?> audioGenerator, StereoAudioRenderer audioRenderer) {
         super(jemu, audioGenerator, audioRenderer);
     }
