+++ src/main/java/io/github/arkosammy12/jemu/frontend/oldui/menus/SettingsMenu.java
@@ -1,6 +1,6 @@
 package io.github.arkosammy12.jemu.frontend.oldui.menus;
 
-import io.github.arkosammy12.jemu.application.Jemu;
+import io.github.arkosammy12.jemu.application.OldJemu;
 import io.github.arkosammy12.jemu.application.io.DataManager;
 import io.github.arkosammy12.jemu.application.io.Serializable;
 import io.github.arkosammy12.jemu.application.io.initializers.ApplicationInitializer;
@@ -24,7 +24,7 @@ public class SettingsMenu extends JMenu implements EmulatorInitializerConsumer {
 
     private final JRadioButtonMenuItem showInfoBarButton;
 
-    public SettingsMenu(Jemu jemu, MainWindow mainWindow) {
+    public SettingsMenu(OldJemu jemu, MainWindow mainWindow) {
         super("Settings");
 
         this.setMnemonic(KeyEvent.VK_S);
